Remove compat module
