--- pykka/_compat/await_py3.py
@@ -1,8 +1,0 @@
-def await_dunder_future(self):
-    yield
-    value = self.get()
-    return value
-
-
-async def await_keyword(val):
-    return await val
